+++ CHANGELOG.md
@@ -6,6 +6,13 @@ Change are listed in reverse chronological order (newest to oldest).
 
 <span class="changelog">
 
+###### [ 1.0.35 ] - 2024/03/18
+
+  * Updated `SpotifyClient.SaveTrackFavorites` method to make the `ids` argument optional; if not supplied, the currently playing id value is used instead.
+  * Updated `SpotifyClient.RemoveTrackFavorites` method to make the `ids` argument optional; if not supplied, the currently playing id value is used instead.
+  * Updated `SpotifyClient.AddPlaylistItems` method to make the `uris` argument optional; if not supplied, the currently playing uri value is used instead.
+  * Updated `SpotifyClient.RemovePlaylistItems` method to make the `uris` argument optional; if not supplied, the currently playing uri value is used instead.
+
 ###### [ 1.0.34 ] - 2024/03/18
 
   * Fixed `SpotifyClient.CreatePlaylist` method to properly pass the Public and Collaborative parameters to the Spotify Web API.  Prior to this, any created playlists were being marked with Public=True.